Fixed build for dronekit-android.3.0.4

+++ Android/src/org/droidplanner/android/fragments/widget/video/WidgetVideoPreferences.kt
@@ -3,7 +3,7 @@ package org.droidplanner.android.fragments.widget.video
 import android.app.DialogFragment
 import android.content.Context
 import android.os.Bundle
-import android.support.annotation.IntDef
+import androidx.annotation.IntDef
 import android.text.TextUtils
 import android.view.LayoutInflater
 import android.view.View
@@ -27,7 +27,7 @@ class WidgetVideoPreferences : DialogFragment() {
         const val CUSTOM_VIDEO_TYPE = 1
     }
 
-    @IntDef(SOLO_VIDEO_TYPE.toLong(), CUSTOM_VIDEO_TYPE.toLong())
+    @IntDef(SOLO_VIDEO_TYPE, CUSTOM_VIDEO_TYPE)
     @Retention(AnnotationRetention.SOURCE)
     annotation class VideoType
